Rename: 절대경로로 변경 (내가헷갈려서)
