+++ src/components/pages/Mypage/components/MyArticleBoard/MyCommentList.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import PreviewArticle from '@components/pages/articles/ArticleListPage/components/PreviewArticle';
 import { Article } from '@root/network/types/Article';
 import { useGetMyComments } from '@components/pages/Mypage/hooks';
-import PageSelector from './PageSelector';
+import PageSelector from '@components/pages/Mypage/components/MyArticleBoard/PageSelector';
 
 const MyCommentList = () => {
   const { comments, maxPage } = useGetMyComments();
