+++ src/components/pages/Mypage/components/MyArticleBoard/PageSelector.tsx
@@ -1,4 +1,4 @@
-import { usePageSelector } from './hooks';
+import { usePageSelector } from '@components/pages/Mypage/hooks';
 
 interface IProps {
   maxPage: number;
